# internal/adapter/http/bridge.gen.go
// Code generated by scripts/bridge.py; DO NOT EDIT.
package http

import (
	"context"
	"encoding/json"
	"github.com/team-screaner/screener-service/api"
	"github.com/team-screaner/screener-service/internal/domain"
	"net/http"
)

type jsonResponse struct{ data domain.Result }

func (r jsonResponse) write(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	return json.NewEncoder(w).Encode(r.data)
}
func (r jsonResponse) VisitRegisterResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) Register(ctx context.Context, request api.RegisterRequestObject) (api.RegisterResponseObject, error) {
	result, err := s.execute(ctx, "register", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitLoginResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) Login(ctx context.Context, request api.LoginRequestObject) (api.LoginResponseObject, error) {
	result, err := s.execute(ctx, "login", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitLogoutResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) Logout(ctx context.Context, request api.LogoutRequestObject) (api.LogoutResponseObject, error) {
	result, err := s.execute(ctx, "logout", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetMeResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetMe(ctx context.Context, request api.GetMeRequestObject) (api.GetMeResponseObject, error) {
	result, err := s.execute(ctx, "getMe", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateTokenResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateToken(ctx context.Context, request api.CreateTokenRequestObject) (api.CreateTokenResponseObject, error) {
	result, err := s.execute(ctx, "createToken", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListTokensResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListTokens(ctx context.Context, request api.ListTokensRequestObject) (api.ListTokensResponseObject, error) {
	result, err := s.execute(ctx, "listTokens", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitRevokeTokenResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) RevokeToken(ctx context.Context, request api.RevokeTokenRequestObject) (api.RevokeTokenResponseObject, error) {
	result, err := s.execute(ctx, "revokeToken", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateOrganizationResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateOrganization(ctx context.Context, request api.CreateOrganizationRequestObject) (api.CreateOrganizationResponseObject, error) {
	result, err := s.execute(ctx, "createOrganization", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListOrganizationsResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListOrganizations(ctx context.Context, request api.ListOrganizationsRequestObject) (api.ListOrganizationsResponseObject, error) {
	result, err := s.execute(ctx, "listOrganizations", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitAddMemberResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) AddMember(ctx context.Context, request api.AddMemberRequestObject) (api.AddMemberResponseObject, error) {
	result, err := s.execute(ctx, "addMember", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListMembersResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListMembers(ctx context.Context, request api.ListMembersRequestObject) (api.ListMembersResponseObject, error) {
	result, err := s.execute(ctx, "listMembers", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateSkillResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateSkill(ctx context.Context, request api.CreateSkillRequestObject) (api.CreateSkillResponseObject, error) {
	result, err := s.execute(ctx, "createSkill", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListSkillsResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListSkills(ctx context.Context, request api.ListSkillsRequestObject) (api.ListSkillsResponseObject, error) {
	result, err := s.execute(ctx, "listSkills", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitDeactivateSkillResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) DeactivateSkill(ctx context.Context, request api.DeactivateSkillRequestObject) (api.DeactivateSkillResponseObject, error) {
	result, err := s.execute(ctx, "deactivateSkill", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListFactTypesResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListFactTypes(ctx context.Context, request api.ListFactTypesRequestObject) (api.ListFactTypesResponseObject, error) {
	result, err := s.execute(ctx, "listFactTypes", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateMatrixResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateMatrix(ctx context.Context, request api.CreateMatrixRequestObject) (api.CreateMatrixResponseObject, error) {
	result, err := s.execute(ctx, "createMatrix", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListMatricesResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListMatrices(ctx context.Context, request api.ListMatricesRequestObject) (api.ListMatricesResponseObject, error) {
	result, err := s.execute(ctx, "listMatrices", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetMatrixResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetMatrix(ctx context.Context, request api.GetMatrixRequestObject) (api.GetMatrixResponseObject, error) {
	result, err := s.execute(ctx, "getMatrix", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateVersionResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateVersion(ctx context.Context, request api.CreateVersionRequestObject) (api.CreateVersionResponseObject, error) {
	result, err := s.execute(ctx, "createVersion", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListVersionsResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListVersions(ctx context.Context, request api.ListVersionsRequestObject) (api.ListVersionsResponseObject, error) {
	result, err := s.execute(ctx, "listVersions", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetVersionResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetVersion(ctx context.Context, request api.GetVersionRequestObject) (api.GetVersionResponseObject, error) {
	result, err := s.execute(ctx, "getVersion", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitUpdateVersionResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) UpdateVersion(ctx context.Context, request api.UpdateVersionRequestObject) (api.UpdateVersionResponseObject, error) {
	result, err := s.execute(ctx, "updateVersion", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitPublishVersionResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) PublishVersion(ctx context.Context, request api.PublishVersionRequestObject) (api.PublishVersionResponseObject, error) {
	result, err := s.execute(ctx, "publishVersion", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitForkMatrixResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ForkMatrix(ctx context.Context, request api.ForkMatrixRequestObject) (api.ForkMatrixResponseObject, error) {
	result, err := s.execute(ctx, "forkMatrix", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitDuplicateMatrixResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) DuplicateMatrix(ctx context.Context, request api.DuplicateMatrixRequestObject) (api.DuplicateMatrixResponseObject, error) {
	result, err := s.execute(ctx, "duplicateMatrix", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetUpstreamResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetUpstream(ctx context.Context, request api.GetUpstreamRequestObject) (api.GetUpstreamResponseObject, error) {
	result, err := s.execute(ctx, "getUpstream", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitReviewUpstreamResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ReviewUpstream(ctx context.Context, request api.ReviewUpstreamRequestObject) (api.ReviewUpstreamResponseObject, error) {
	result, err := s.execute(ctx, "reviewUpstream", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateUserMatrixResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateUserMatrix(ctx context.Context, request api.CreateUserMatrixRequestObject) (api.CreateUserMatrixResponseObject, error) {
	result, err := s.execute(ctx, "createUserMatrix", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListUserMatricesResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListUserMatrices(ctx context.Context, request api.ListUserMatricesRequestObject) (api.ListUserMatricesResponseObject, error) {
	result, err := s.execute(ctx, "listUserMatrices", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitSetOverrideResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) SetOverride(ctx context.Context, request api.SetOverrideRequestObject) (api.SetOverrideResponseObject, error) {
	result, err := s.execute(ctx, "setOverride", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateEvidenceResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateEvidence(ctx context.Context, request api.CreateEvidenceRequestObject) (api.CreateEvidenceResponseObject, error) {
	result, err := s.execute(ctx, "createEvidence", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListEvidenceResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListEvidence(ctx context.Context, request api.ListEvidenceRequestObject) (api.ListEvidenceResponseObject, error) {
	result, err := s.execute(ctx, "listEvidence", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitBatchEvidenceResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) BatchEvidence(ctx context.Context, request api.BatchEvidenceRequestObject) (api.BatchEvidenceResponseObject, error) {
	result, err := s.execute(ctx, "batchEvidence", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitReviewEvidenceResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ReviewEvidence(ctx context.Context, request api.ReviewEvidenceRequestObject) (api.ReviewEvidenceResponseObject, error) {
	result, err := s.execute(ctx, "reviewEvidence", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateAssessmentResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateAssessment(ctx context.Context, request api.CreateAssessmentRequestObject) (api.CreateAssessmentResponseObject, error) {
	result, err := s.execute(ctx, "createAssessment", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListAssessmentsResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListAssessments(ctx context.Context, request api.ListAssessmentsRequestObject) (api.ListAssessmentsResponseObject, error) {
	result, err := s.execute(ctx, "listAssessments", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetAssessmentResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetAssessment(ctx context.Context, request api.GetAssessmentRequestObject) (api.GetAssessmentResponseObject, error) {
	result, err := s.execute(ctx, "getAssessment", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetGrowthContextResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetGrowthContext(ctx context.Context, request api.GetGrowthContextRequestObject) (api.GetGrowthContextResponseObject, error) {
	result, err := s.execute(ctx, "getGrowthContext", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetGapsResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetGaps(ctx context.Context, request api.GetGapsRequestObject) (api.GetGapsResponseObject, error) {
	result, err := s.execute(ctx, "getGaps", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitGetRadarResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) GetRadar(ctx context.Context, request api.GetRadarRequestObject) (api.GetRadarResponseObject, error) {
	result, err := s.execute(ctx, "getRadar", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitCreateGrowthPlanResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) CreateGrowthPlan(ctx context.Context, request api.CreateGrowthPlanRequestObject) (api.CreateGrowthPlanResponseObject, error) {
	result, err := s.execute(ctx, "createGrowthPlan", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListGrowthPlansResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListGrowthPlans(ctx context.Context, request api.ListGrowthPlansRequestObject) (api.ListGrowthPlansResponseObject, error) {
	result, err := s.execute(ctx, "listGrowthPlans", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitUpdateGrowthPlanResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) UpdateGrowthPlan(ctx context.Context, request api.UpdateGrowthPlanRequestObject) (api.UpdateGrowthPlanResponseObject, error) {
	result, err := s.execute(ctx, "updateGrowthPlan", request.Id)
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitImportXLSXResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ImportXLSX(ctx context.Context, request api.ImportXLSXRequestObject) (api.ImportXLSXResponseObject, error) {
	result, err := s.execute(ctx, "importXLSX", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}
func (r jsonResponse) VisitListAuditResponse(w http.ResponseWriter) error { return r.write(w) }
func (s *server) ListAudit(ctx context.Context, request api.ListAuditRequestObject) (api.ListAuditResponseObject, error) {
	result, err := s.execute(ctx, "listAudit", "")
	if err != nil {
		return nil, err
	}
	return jsonResponse{data: result}, nil
}